Port cartographer-project/async_grpc to gRPC+flatbuffers
Make the async_grpc framework support flatbuffer gRPC. This required
quite a few fundamental changes to the internals of async_grpc that it
didn't seem practical to make this part of the configuration, so I'm
just forking the library.

This patch is kind of huge because with flatbuffers::grpc, the message
type does not have a common base class. So many functions in Rpc and
RpcHandler needed to be changed to deal with a concrete message
type. This is unlike typical gRPC where all messages inherit from
google::protobuf::Message.

This means that Rpc needed to be turned into a template. This ended up
working fine though since Rpc owns the handler and the types are known
at construction time.

Some additional changes were made to break the dependency cycle
between Rpc and Service. These are no longer friend classes. Instead a
ServiceInterface was extracted from Service which exposes the
underlying grpc::Service API.

After these refactorings RpcHandlerInterface was no longer
needed. This is because RpcHandlerInfo doesn't need to have a factory
for constructing the Handler, since the Rpc template does it, and that
actually knows the Request and Response type.

Defined Hunter package info. If you intend to use this without Hunter
then just set the CMake variable HUNTER_ENABLED=OFF and this will just
be a typical CMake project where the system is searched for
dependencies.

Updated README.md.

clang-format was run on everything which picked up a few formatting
inconsistencies from the original async_grpc code, but nothing major.